virgildiana_telegram_client/requirements.txt
python-telegram-bot==13.7
urllib3<2
requests
